# noblelift-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Noblelift App API
  version: 1.0.0
  description: |
    REST API для корпоративной системы Noblelift.
    Согласовано с ТЗ и текущим фронтендом. Таймстемпы — Unix ms (int64), camelCase в JSON.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: Список авто
      parameters:
        - in: query
          name: isActive
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedVehicles' }
    post:
      tags: [Vehicles]
      summary: Добавить авто (super_admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
  /vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: Карточка авто (с текущим держателем)
      parameters: [ { $ref: '#/components/parameters/VehicleId' } ]
      responses:
        '200':
          description: Vehicle
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VehicleDetails' }
    patch:
      tags: [Vehicles]
      summary: Обновить авто
      parameters: [ { $ref: '#/components/parameters/VehicleId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleUpdate' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
    delete:
      tags: [Vehicles]
      summary: Исключить из парка
      parameters: [ { $ref: '#/components/parameters/VehicleId' } ]
      responses:
        '204': { description: No Content }
  /vehicles/{id}/logs:
    get:
      tags: [Vehicles]
      summary: Журнал действий по авто
      parameters:
        - { $ref: '#/components/parameters/VehicleId' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Logs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedVehicleLogs' }
  /vehicles/{id}/take:
    post:
      tags: [Vehicles]
      summary: Взять авто
      parameters: [ { $ref: '#/components/parameters/VehicleId' } ]
      responses:
        '201':
          description: Created (log)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VehicleLog' }
  /vehicles/{id}/release:
    post:
      tags: [Vehicles]
      summary: Оставить авто
      parameters: [ { $ref: '#/components/parameters/VehicleId' } ]
      responses:
        '201':
          description: Created (log)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VehicleLog' }

  # ---------- Audit ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    Offset:
      in: query
      name: offset
      schema: { type: integer, default: 0, minimum: 0 }
...
    VehicleId:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    Vehicle:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        plate: { type: string }
        color: { type: string, nullable: true }
        isActive: { type: boolean }
    VehicleCreate:
      type: object
      required: [name, plate]
      properties:
        name: { type: string }
        plate: { type: string }
        color: { type: string, nullable: true }
        isActive: { type: boolean, default: true }
    VehicleUpdate:
      type: object
      properties:
        name: { type: string }
        plate: { type: string }
        color: { type: string, nullable: true }
        isActive: { type: boolean }
    VehicleLogAction:
      type: string
      enum: [take, release]
    VehicleLog:
      type: object
      properties:
        id: { type: integer, format: int64 }
        vehicleId: { type: integer, format: int64 }
        userId: { type: integer, format: int64 }
        action: { $ref: '#/components/schemas/VehicleLogAction' }
        createdAt: { type: integer, format: int64 }
        note: { type: string, nullable: true }
    VehicleDetails:
      type: object
      properties:
        vehicle: { $ref: '#/components/schemas/Vehicle' }
        currentHolder:
          type: object
          nullable: true
          properties:
            userId: { type: integer, format: int64 }
            fullName: { type: string }
            takenAt: { type: integer, format: int64 }
    PaginatedVehicles:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Vehicle' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    PaginatedVehicleLogs:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/VehicleLog' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }

    # ----- Audit -----
...